Prime generation implemented
This is basically a really stupid random algorithm and another shot of
some CRT-based idea which I had (that, unfortunately, is slower than the
traditional method). I might try re-reading the fastprimes paper again
and try to understand it (we're not looking for security, just speed in
this case). Note that generated primes are not cryptographically secure
by any stretch of the imagination, so do not use these for anything but
pentesting.