Reworking of code to prepare for hardware tokens
When we want to handle private keys that are not on files, but only
indirecly accessible (for example through as HSM such as the NitroKey)
then we cannot use all traditional code to access those keys (because
they're simply not there). Instead, we need to go through the OpenSSL
engine. For this, some rework was necessary of all the code, because we
now always generate OpenSSL configuration files on-the-fly. This
additionally makes for a lot cleaner code. Also added numerous testcases
for the OpenSSLTools facility.