Show a nice error message when certificate creation fails
When trying to create certificates with a duplicate common name, the
error message is kind of cryptic (OpenSSL fails and, when x509sak is
executed verbosely, the error is 'failed to update database' followed by
'TXT_DB error number 2'). Because this is a fairly common error, show a
nice error message by default unless more verbosity is requested.