Make PKCS#11 module configurable and add PKCS#11 tests
The shared object that's used for PKCS#11 access is now configurable and
the .so search path is as well. This opens up the possibility to have
HSM tests as well, which is pretty awesome: We use SoftHSM2 to create a
software-only PKCS#11 token and run some CA creation/certificate signing
scenarios with it.