Fixed build failure on Travis-CI
Previously, stderr was always merged with stdout. In the recent
SubprocessExecutor, we fixed that. More recent versions of OpenSSL
output the depth/failure message of "openssl verify" on stderr and so
locally our builds worked with only parsing the output of stderr.
However, on the older OpenSSL version on Travis-CI, stdout is used for
this message (therefore finding no regex match and thinking that cert is
not signed by issuer). Fixed by always looking at stdout and stderr.