Implemented "hashpart" command
During ongoing implementation of the EdDSA code, I needed a way to brute
force hashes inside binary blobs. Since I've implemented this very thing
a couple of times before, I thought it would be nice to have this
properly as part of the toolchain. It has been added as the "hashpart"
facility. Originally there was multithreading support, but using
multiple threads significantly slowed down computation (GIL issue?). So
in this commit, it's completely removed again.

Note that this has been split off into two commits artificially and
after the fact that EdDSA was refactored. So there's some minor overlap
in __main__.py code, causing this commit by itself not to work -- it
will fail many tests. However, I can't be bothered to retrofit this and
will commit it like it is.