Refactored test file access
Instead of reading the files directly and assuming it'll be found under
the x509sak/tests/data directory, we now read it through
pkgutil.get_data() and then write it to a NamedTemporaryFile using the
ResourceFileLoader context handler. This should work eventually (with
some minor modifications) even if the whole package is packed up in a
egg.