Use UTF-8 by default when calling OpenSSL
We assume that all strings that are input are UTF-8. It's 2018, after
all, and we're not supporting any other shitty encoding. Note that on
UTF-8 system, this fixes a bug where characters were encoded wrong
(e.g., "ö" would have become interpreted as the latin1 \xc3\xb6 which
then would decode to "Ã¶").