Allow "examinecert" action to read JSON, not write it
Pretty significant refactoring of the "examinecert" guts; first, we
unified the output format (instead of printing on stdout and having an
optional JSON sink as well) - we can now combine any output format
(text, text w/ ANSI colors, JSON) with any output sink (stdout or file).
Secondly, refactor the ActionExamineCert to not handle the internal data
(e.g., enums before they're serialized into a JSON-friendly format), but
only print what's also present in the JSON itself. This then allows for
easy reading of the JSON output in order to re-display it. One useful
application is to scan a certificate and output the scan to the (raw)
JSON, then re-print it to ANSI text. Or maybe we also add an HTML
facility later on as well.