Also judge length of salt for RSA/PSS
Length of salt is also judged now for RSAPSS. Some modifications were
needed to always show the topic of the text when doing bit estimations.
Otherwise it would, for instance, show "Security: exceptionally high
level of security" even when only a very long salt was used in RSAPSS.
We want to make it clear that judgement only applies to the used salt.