Fix bug with SAN directoryName and refactor SecurityEstimator
There was an error with directoryName entries for the SAN because we
couldn't properly represent them as strings. Fix this properly and
decode them as DNs. Now we had a cyclic dependency because the DN class
invoked SecurityEstimator in the anylyze() method. To fix all this,
refactor all code to not have a .analyze() method (i.e., have all
dependencies one-way).