Added ability to discern malformed RSA/PSS parameters
When the parameter field is NULL, we previously got an ASN.1 exception.
Catch it and instead raise the proper error code. In a similar vein, I
cannot think of a way to produce a testcas that has trailing ASN.1 data
which does not result in a wrong encoding (i.e., directly leads to
schema violation). Need to think of something here (either remove the
judgement code or produce a testcase).