Rework calling x509sak as subprocess
For running code coverage checks, we need to be able to set the
COVERAGE_FILE environment because some of these tests change the working
directory and this messes things up otherwise (.coverage is lost). So
rework the whole way we call subprocesses to be able to inject that
environment variable if coverage checks are requested.