Added more tests around DNs, BijectiveDict and CRLDP
We would like to also evaluate the DN that might be embedded in an LDAP
URI inside a GeneralName. For this we had to adapt the DistinguishedName
class to allow string parsing, something that wasn't needed before. Also
in URIs we frequently see "cn=" instead of "CN=", i.e., we need
case-insensitive lookup of the OIDs. For this we modified the
BijectiveDict to allow key and value predicates that allow arbitrary
matching. Then we added the testcases we liked to have in CRL DPs, but
also have unit tests for the new BijectiveDict and DN functionality.