Some more work on NameConstraints
While I've been doing some more work on the NameConstraints, I feel this
isn not the correct way forward to classify everything. I need to
re-think the whole classification scheme and switch to something more
scalable and robust. We currently have ~300 error codes and many of
these are duplicates of each other -- for example, GeneralName
validation errors share a large portion of their code, yet nothing of
that is reused. I like having explicit code points because it makes
nothing fall under the radar. However, it would be really nice to have a
specific field of a specific extension and just be able to say "import
all GeneralName validation errors" that could apply. Therefore, I'll now
focus on that part as the next step before moving further forward.